+++ source/renderer/app/containers/staking/dialogs/DelegationSetupWizardDialogContainer.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
-import { map } from 'lodash';
+import { map, find } from 'lodash';
 import DelegationSetupWizardDialog from '../../../components/staking/delegation-setup-wizard/DelegationSetupWizardDialog';
 import { formattedWalletAmount } from '../../../utils/formatters';
 import { MIN_DELEGATION_FUNDS } from '../../../config/stakingConfig';
@@ -39,9 +39,15 @@ const messages = defineMessages({
 
 type State = {
   activeStep: number,
+  selectedWalletId: ?string,
 };
 
 type Props = InjectedDialogContainerProps;
+
+const initialState = {
+  activeStep: 0,
+  selectedWalletId: null,
+};
 
 @inject('stores', 'actions')
 @observer
@@ -61,7 +67,7 @@ export default class DelegationSetupWizardDialogContainer extends Component<
   };
 
   state = {
-    activeStep: 0,
+    ...initialState,
   };
 
   STEPS_LIST = [
@@ -72,17 +78,12 @@ export default class DelegationSetupWizardDialogContainer extends Component<
   ];
 
   handleDialogClose = () => {
+    this.setState({ ...initialState });
     this.props.actions.dialogs.closeActiveDialog.trigger();
   };
 
   handleContinue = () => {
     const { activeStep } = this.state;
-
-    // TODO - remove once all steps are constructed
-    if (activeStep === 1) {
-      this.handleDialogClose();
-    }
-
     this.setState({ activeStep: activeStep + 1 });
   };
 
@@ -98,10 +99,26 @@ export default class DelegationSetupWizardDialogContainer extends Component<
     this.props.stores.app.openExternalLink(learnMoreLinkUrl);
   };
 
+  handleConfirm = () => {
+    // @TODO - proceed confirmation data
+    this.handleContinue();
+  };
+
+  handleActivate = () => {
+    // @TODO - proceed activation data
+    this.handleDialogClose();
+  };
+
+  handleSelectWallet = (walletId: string) => {
+    this.setState({ selectedWalletId: walletId });
+    this.handleContinue();
+  };
+
   render() {
-    const { activeStep } = this.state;
-    const { stores } = this.props;
-    const { wallets } = stores;
+    const { activeStep, selectedWalletId } = this.state;
+    const { app, staking, wallets, profile } = this.props.stores;
+    const { currentTheme } = profile;
+    const { stakePools, delegatingStakePools } = staking;
 
     let setupDisabled = true;
     const walletsData = map(wallets.all, wallet => {
@@ -114,11 +131,18 @@ export default class DelegationSetupWizardDialogContainer extends Component<
       }
 
       return {
-        value,
+        id: wallet.id,
         label: wallet.name,
+        value,
         isAcceptableSetupWallet,
+        hasPassword: wallet.hasPassword,
       };
     });
+
+    const selectedWallet = find(
+      walletsData,
+      wallet => wallet.id === selectedWalletId
+    );
 
     return (
       <DelegationSetupWizardDialog
@@ -127,10 +151,18 @@ export default class DelegationSetupWizardDialogContainer extends Component<
         activeStep={activeStep}
         minDelegationFunds={MIN_DELEGATION_FUNDS}
         isDisabled={activeStep === 1 && setupDisabled}
+        selectedWallet={selectedWallet || null}
+        stakePoolsList={stakePools}
+        stakePoolsDelegatingList={delegatingStakePools}
+        onOpenExternalLink={app.openExternalLink}
+        currentTheme={currentTheme}
         onClose={this.handleDialogClose}
         onContinue={this.handleContinue}
+        onSelectWallet={this.handleSelectWallet}
         onBack={this.onBack}
         onLearnMoreClick={this.handleLearnMoreClick}
+        onConfirm={this.handleConfirm}
+        onActivate={this.handleActivate}
       />
     );
   }