+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -238,6 +238,8 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-block-consolidation-stripe-light-2-background-color': '#fafbfc',
     '--theme-block-consolidation-button-background-color': '#243e62',
     '--theme-block-consolidation-button-background-color-hover': '#fafbfc',
+    '--theme-block-consolidation-button-icon-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-block-consolidation-button-icon-color-hover': '#243e62',
     '--theme-block-consolidation-button-text-color-hover': '#243e62',
     '--theme-block-consolidation-button-border-color': '#fafbfc',
   },
@@ -307,6 +309,41 @@ const LIGHT_BLUE_THEME_CONFIG = {
       'rgba(250, 251, 252, 0.5)',
   },
   delegationSetupWizard: {
+    '--theme-delegation-steps-activation-steps-indicator-color': '#5e6066',
+    '--theme-delegation-steps-activation-description-color':
+      'rgba(94, 96, 102, 0.8)',
+    '--theme-delegation-steps-activation-fees-label-color': '#5e6066',
+    '--theme-delegation-steps-activation-fees-amount-color': '#ea4c5b',
+    '--theme-delegation-steps-activation-address-value-color': '#5e6066',
+    '--theme-delegation-steps-choose-stake-pool-checkmark-icon-color':
+      '#c6cdd6',
+    '--theme-delegation-steps-choose-stake-pool-delegated-pools-label-color':
+      'rgba(94, 96, 102, 1)',
+    '--theme-delegation-steps-choose-stake-pool-slug-color': '#c6cdd6',
+    '--theme-delegation-steps-choose-stake-pool-select-box-placeholder-color':
+      '#c6cdd6',
+    '--theme-delegation-steps-choose-stake-pool-selected-checkmark-icon-color':
+      '#fafbfc',
+    '--theme-delegation-steps-choose-stake-pool-selected-slug-color': '#fafbfc',
+    '--theme-delegation-steps-choose-stake-pool-title-color':
+      'rgba(94, 96, 102, 0.8)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-arrow-color':
+      'rgba(94, 96, 102, 0.9)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-background-color':
+      'rgba(94, 96, 102, 0.9)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-box-shadow':
+      '0 5px 20px 0 rgba(0, 0, 0, 0.25)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-description-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-slug-color':
+      'rgba(250, 251, 252, 0.6)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-table-label-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-table-value-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-title-color':
+      'rgba(250, 251, 252, 1)',
+    '--theme-delegation-steps-choose-stake-pool-tooltip-url-color': '#85b6f9',
     '--theme-delegation-steps-choose-wallet-custom-value-color': '#5e6066',
     '--theme-delegation-steps-choose-wallet-description-color':
       'rgba(94, 96, 102, 0.8)',
@@ -319,6 +356,11 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-delegation-steps-choose-wallet-error-select-options-color':
       '#5e6066',
     '--theme-delegation-steps-choose-wallet-steps-indicator-color': '#5e6066',
+    '--theme-delegation-steps-confirmation-description-color':
+      'rgba(94, 96, 102, 0.8)',
+    '--theme-delegation-steps-confirmation-fees-label-color': '#5e6066',
+    '--theme-delegation-steps-confirmation-fees-amount-color': '#ea4c5b',
+    '--theme-delegation-steps-confirmation-steps-indicator-color': '#5e6066',
     '--theme-delegation-steps-intro-content-text-color':
       'rgba(94, 96, 102, 0.8)',
     '--theme-delegation-steps-intro-divider-border-color': '#dfe4e8',
@@ -397,6 +439,9 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-input-right-floating-text-color': 'rgba(94, 96, 102, 0.5)',
     '--theme-input-text-color': '#5e6066',
   },
+  link: {
+    '--theme-link-main-color': 'rgba(41, 111, 208, 1)',
+  },
   loading: {
     '--theme-loading-background-color': '#fafbfc',
     '--theme-loading-no-disk-space-background-color': 'rgba(171, 23, 0, 0.94)',
@@ -414,6 +459,8 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-manual-update-overlay-button-background-color': '#243e62',
     '--theme-manual-update-overlay-button-background-color-hover': '#fafbfc',
     '--theme-manual-update-overlay-button-border-color': '#fafbfc',
+    '--theme-manual-update-overlay-button-icon-color': '#fafbfc',
+    '--theme-manual-update-overlay-button-icon-color-hover': '#243e62',
     '--theme-manual-update-overlay-button-text-color-hover': '#243e62',
     '--theme-manual-update-overlay-text-color': 'rgba(250, 251, 252, 0.7)',
     '--theme-manual-update-overlay-text-highlight-color': '#fafbfc',
@@ -474,7 +521,7 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-node-update-button-text-color': '#fafbfc',
   },
   notification: {
-    '--theme-notification-message-background-color': 'rgba(36, 61, 98, 0.88)',
+    '--theme-notification-message-background-color': 'rgba(36, 61, 98, 0.95)',
     '--theme-notification-message-text-color': '#fafbfc',
     '--theme-legacy-badge-background-color': '#ab1700',
     '--theme-legacy-notification-background-color': '#ab2712',
@@ -518,6 +565,7 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-report-issue-connecting-background-color':
       'rgba(250, 251, 252, 0.05)',
     '--theme-report-issue-connecting-text-color': '#fafbfc',
+    '--theme-report-issue-icon-color': '#fafbfc',
     '--theme-report-issue-syncing-background-color': 'rgba(94, 96, 102, 0.05)',
     '--theme-report-issue-syncing-text-color': '#5e6066',
     '--theme-report-issue-syncing-download-logs-text-color': '#5e6066',
@@ -559,38 +607,41 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-sidebar-menu-add-button-text-color': '#fafbfc',
   },
   stakePools: {
-    '--theme-staking-stake-pools-title-color': '#5e6066',
-    '--theme-staking-stake-pools-search-button-color': '#5e6066',
     '--theme-staking-stake-pool-background-color': '#fafbfc',
     '--theme-staking-stake-pool-border-color': '#c6cdd6',
     '--theme-staking-stake-pool-glow-color': 'rgba(0, 149, 255, 0.3)',
-    '--theme-staking-stake-pool-slug-color': '#5e6066',
     '--theme-staking-stake-pool-retirement-background-color': '#ea4c5b',
+    '--theme-staking-stake-pool-selected-background-color': '#5da377',
+    '--theme-staking-stake-pool-selected-checkmark-icon-color': '#fafbfc',
+    '--theme-staking-stake-pool-selected-slug-color': '#fafbfc',
+    '--theme-staking-stake-pool-slug-color': '#5e6066',
     '--theme-staking-stake-pool-tooltip-background-color': `${chroma(
       '#FFFFFF'
     ).alpha(0.97)}`,
     '--theme-staking-stake-pool-tooltip-border-color': '#c6cdd6',
-    '--theme-staking-stake-pool-tooltip-shadow-color': 'rgba(0, 0, 0, 0.25)',
-    '--theme-staking-stake-pool-tooltip-text-color': '#5e6066',
-    '--theme-staking-stake-pool-tooltip-link-color': '#296fd0',
-    '--theme-staking-stake-pool-tooltip-table-title-color': '#5e6066',
-    '--theme-staking-stake-pool-tooltip-table-param-color': '#5e6066',
-    '--theme-staking-stake-pool-tooltip-retirement-text-color': '#fafbfc',
-    '--theme-staking-stake-pool-tooltip-retirement-background-color': '#ea4c5b',
-    '--theme-staking-stake-pool-tooltip-delegate-button-background-color':
-      '#2f496e',
-    '--theme-staking-stake-pool-tooltip-delegate-button-hover-background-color':
-      '#445b7c',
     '--theme-staking-stake-pool-tooltip-delegate-button-active-background-color':
       '#1e304a',
-    '--theme-staking-stake-pool-tooltip-delegate-button-text-color': '#fafbfc',
-    '--theme-staking-stake-pool-tooltip-delegate-button-inverse-text-color':
-      '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-delegate-button-background-color':
+      '#2f496e',
     '--theme-staking-stake-pool-tooltip-delegate-button-border-color':
       'transparent',
+    '--theme-staking-stake-pool-tooltip-delegate-button-hover-background-color':
+      '#445b7c',
+    '--theme-staking-stake-pool-tooltip-delegate-button-inverse-text-color':
+      '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-delegate-button-text-color': '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-link-color': '#296fd0',
+    '--theme-staking-stake-pool-tooltip-retirement-background-color': '#ea4c5b',
+    '--theme-staking-stake-pool-tooltip-retirement-text-color': '#fafbfc',
+    '--theme-staking-stake-pool-tooltip-shadow-color': 'rgba(0, 0, 0, 0.25)',
+    '--theme-staking-stake-pool-tooltip-table-param-color': '#5e6066',
+    '--theme-staking-stake-pool-tooltip-table-title-color': '#5e6066',
+    '--theme-staking-stake-pool-tooltip-text-color': '#5e6066',
+    '--theme-staking-stake-pools-search-button-color': '#5e6066',
+    '--theme-staking-stake-pools-search-icon-color': 'rgba(94, 96, 102, 1)',
+    '--theme-staking-stake-pools-title-color': '#5e6066',
   },
   staking: {
-    '--theme-staking-background-color': 'rgba(52, 70, 94, 0.1)',
     '--theme-staking-content-background-color': '#fafbfc',
     '--theme-staking-content-border-color': '#c6cdd6',
     '--theme-staking-font-color-accent': '#5e6066',
@@ -601,6 +652,7 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-staking-table-border-color': '#c6cdd6',
     '--theme-staking-link-color': 'rgba(41, 111, 208, 1)',
     '--theme-staking-link-color-light': 'rgba(41, 111, 208, 0.5)',
+    '--theme-staking-progress-bar-background-color': 'rgba(52, 70, 94, 0.1)',
     '--theme-staking-progress-stripe-dark-1-background-color': '#34465e',
     '--theme-staking-progress-stripe-dark-2-background-color': '#445b7c',
     '--theme-staking-table-body-highlighted-text-color': '#296fd0',
@@ -613,6 +665,10 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-staking-delegation-center-gear-icon-fill-color':
       'rgba(94, 96, 102, 0.5)',
     '--theme-staking-delegation-center-gear-icon-fill-color-active': '#5e6066',
+    '--theme-staking-info-learn-more-button-color': '#fafbfc',
+    '--theme-staking-info-learn-more-icon-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-staking-learn-more-button-color': '#fafbfc',
+    '--theme-staking-learn-more-icon-color': 'rgba(250, 251, 252, 0.7)',
   },
   support: {
     '--theme-support-settings-item-color': 'rgba(94, 96, 102, 0.5)',
@@ -626,6 +682,7 @@ const LIGHT_BLUE_THEME_CONFIG = {
   systemError: {
     '--theme-system-error-overlay-attention-icon-color': '#fafbfc',
     '--theme-system-error-overlay-background-color': 'rgba(171, 23, 0, 0.94)',
+    '--theme-system-error-overlay-support-link-icon-color': '#fafbfc',
     '--theme-system-error-overlay-text-color': '#fafbfc',
   },
   tabs: {